split stories into separate files
